display aria2 launch cmd that doesn't crash
The program crashes because it tries to use mp_downloader->getAria2LaunchCmd()
but the mp_downloader is null if the aria2 launch command fail.

Now the program displays an error message to warn users that all download fct will not working.

This PR https://github.com/kiwix/kiwix-lib/pull/306 is needed in order to display the aria2 launch cmd